third_party/blink/web_tests/http/tests/performance-timing/element-timing/progressively-loaded-image.html

<!DOCTYPE html>
<html>
<meta charset=utf-8 />
<title>Element Timing: buffer elements before onload</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src="resources/element-timing-helpers.js"></script>
<body>
<img src=/resources/slow-image.php?name=square20.png&mimeType=image&sleep=500>
<script>
  let beforeRender;
  // Number of bytes to be read on the initial read, before sleeping.
  // Should be sufficient to do at least a first scan.
  let numInitial = 300;
  let sleep = 500;
  async_test(function(t) {
    const observer = new PerformanceObserver(
      t.step_func_done(function(entryList) {
        assert_equals(entryList.getEntries().length, 1);
        const entry = entryList.getEntries()[0];
        // Since the image is only fully loaded after the sleep, the render timestamp
        // must be greater than |beforeRender| + |sleep|.
        checkElement(entry, 'my_image', beforeRender + sleep);
      })
    );
    observer.observe({entryTypes: ['element']});

    const img = document.createElement('img');
    img.src = '/resources/progressive-image.php?name=square20.jpg&numInitial='
      + numInitial + '&sleep=' + sleep;
    img.setAttribute('elementtiming', 'my_image');
    document.body.appendChild(img)
    beforeRender = performance.now();
  }, "Element Timing: image render timestamp occurs after it is fully loaded.");
</script>
